split test_runtime.py into multiple tests in new tests/runtime folder
